pyproject.toml
[project]
name = "rising-wave"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "confluent-kafka",
    "psycopg2-binary",
]
